chap23문제 파일 위치 변경

+++ src/ch23/chap23_2.java
@@ -6,7 +6,7 @@ import java.util.TreeSet;
 class ComparatorIntegerDemo implements Comparator<Integer> {
     @Override
     public int compare(Integer o1, Integer o2) {
-        return o2.intValue() - o1.intValue();
+        return o2 - o1;
     }
 }
 public class chap23_2 {
